// front/project/src/open_api/models/User.ts
/* tslint:disable */
/* eslint-disable */
/**
 * OpenAPI Tutorial
 * OpenAPI Tutorial by halhorn
 *
 * The version of the OpenAPI document: 0.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { exists, mapValues } from '../runtime';
/**
 * 会員
 * @export
 * @interface User
 */
export interface User {
    /**
     * id
     * @type {number}
     * @memberof User
     */
    id?: number;
    /**
     * 氏名
     * @type {string}
     * @memberof User
     */
    name?: string;
    /**
     * メールアドレス
     * @type {string}
     * @memberof User
     */
    email?: string;
    /**
     * パスワード
     * @type {string}
     * @memberof User
     */
    password?: string;
    /**
     * 自己PR
     * @type {string}
     * @memberof User
     */
    selfPr?: string;
    /**
     * 電話番号
     * @type {string}
     * @memberof User
     */
    tel?: string;
    /**
     * 更新日時
     * @type {Date}
     * @memberof User
     */
    updatedAt?: Date;
    /**
     * 作成日時
     * @type {Date}
     * @memberof User
     */
    createdAt?: Date;
}

export function UserFromJSON(json: any): User {
    return UserFromJSONTyped(json, false);
}

export function UserFromJSONTyped(json: any, ignoreDiscriminator: boolean): User {
    if ((json === undefined) || (json === null)) {
        return json;
    }
    return {
        
        'id': !exists(json, 'id') ? undefined : json['id'],
        'name': !exists(json, 'name') ? undefined : json['name'],
        'email': !exists(json, 'email') ? undefined : json['email'],
        'password': !exists(json, 'password') ? undefined : json['password'],
        'selfPr': !exists(json, 'selfPr') ? undefined : json['selfPr'],
        'tel': !exists(json, 'tel') ? undefined : json['tel'],
        'updatedAt': !exists(json, 'updatedAt') ? undefined : (new Date(json['updatedAt'])),
        'createdAt': !exists(json, 'createdAt') ? undefined : (new Date(json['createdAt'])),
    };
}

export function UserToJSON(value?: User | null): any {
    if (value === undefined) {
        return undefined;
    }
    if (value === null) {
        return null;
    }
    return {
        
        'id': value.id,
        'name': value.name,
        'email': value.email,
        'password': value.password,
        'selfPr': value.selfPr,
        'tel': value.tel,
        'updatedAt': value.updatedAt === undefined ? undefined : (value.updatedAt.toISOString().substr(0,10)),
        'createdAt': value.createdAt === undefined ? undefined : (value.createdAt.toISOString().substr(0,10)),
    };
}